+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [7.14.9] - (xx.06.2024)
+## [7.15.0] - (08.06.2024)
+
+## Changes
+
+- added encharge backup level sensors
+- readme updated
+- config schema updated
+- cleanup
+
+## [7.14.9] - (08.06.2024)
 
 ## Changes